Fill in pixel gaps in drive bar
The devil's in the details, and this particular devil - rounding a CSS
percentage length to the nearest pixel - happens to be inconsistent over
browsers. You could untangle this Gordian knot by manually rounding with
every resize of the window – or just round it up.

(Subpixels! Never get into web development, kids)